Add a public key to your GitHub account

gh_store_key() adds the public key associated with a key ID to your GitHub account. If you have a GitHub Personal Access Token it will attempt to use it; if it fails, it will print the public key for you to copy manually into GitHub.

gh_store_key(key, .token = NULL)

Arguments

key
A character string containing the ID of a key to use. See gpg::gpg_list_keys(); if you haven't created a key, see sign_commits_with_key() or gpg::gpg_keygen().

.token
GitHub Personal Access Token with at least write:gpg_key scope enabled. You can grant access to tokens here.

Details

If you do not have a GitHub Personal Access Token setup or you want to store your key on Gitlab or other service you can either call this function without a token and then add the printed public key manually or call gpg::gpg_export() with newkey and add the returned public key manually. See https://help.github.com/articles/creating-a-personal-access-token-for-the-command-line/ for more information on tokens.

Examples

# NOT RUN {
newkey <- sign_commits_with_key("John Doe", "johndoe@example.com")
# if you do not have personal access token for github
gh_store_key(newkey)
# if your GitHub Personal Access Token is stored in `.Renviron` as GITHUB_PAT
gh_store_key(newkey, Sys.getenv('GITHUB_PAT'))
# }