Add a public key to your GitHub or gitlab or other account

store_public_key() adds the public key associated with a key ID to your GitHub or Gitlab account if token is available. Otherwise it will print the public key for you to copy manually into GitHub/Gitlab or elsewhere.

store_public_key(key, service = NULL, .token = NULL,
  gitlab_url = NULL, open_url = is_interactive())

Arguments

key
A character string containing the ID of a key to use, use the return value of sign_commits_with_key(). If you are not sure, what key you want to use, you can list your locally available keys with gpg::gpg_list_keys(). The email address corresponding to this key must be identical to the email address that you use with GitHub/Gitlab etc.

service
NULL currently supported services: "gh" for GitHub (access token is necessary) or "gl" for gitlab (url and access token is necessary). If left NULL, no upload is attempted, public key is returned for manual upload.

.token
Either GitHub Personal Access Token with at least write:gpg_key scope enabled. You can grant access to tokens here. Supply this with service = "gh". Or Gitlab personal access token with at least api scope enabled. Supply this with service = "gl". If your PAT is stored in an environment variable GITHUB_PAT (for service = "gh") or GITLAB_PAT (for service = "gl") you do not need to supply the token manually.

gitlab_url
NULL by default, or your gitlab url for service = "gl".

open_url
logical, whether open relevant URLs automatically.

Details

If you do not have a Personal Access Token set up or you want to store your key on other service you can either call this function without a token and then add the printed public key manually or call cat(gpg::gpg_export()) with newkey and add the returned public key manually. For more information see GitHub tokens or gitlab tokens.

Examples

# NOT RUN {
new_key <- sign_commits_with_key("John Doe", "johndoe@example.com")
# if you do not have personal access token for github or gitlab
store_public_key(key = new_key)
# if your GitHub Personal Access Token is stored in `.Renviron` as MY_GITHUB_PAT
store_public_key(key = new_key, service = "gh", .token = Sys.getenv('MY_GITHUB_PAT'))
# }